Rename package rabbitmq to amqp
